add Push class to control push notification
